refactor: Session Manager를 활용하도록 LoginController 리팩터링
