+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -2,10 +2,13 @@ package com.techcourse.controller;
 
 import java.io.IOException;
 import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
 
+import org.apache.catalina.manager.SessionManager;
 import org.apache.coyote.http11.HttpCookie;
 import org.apache.coyote.http11.HttpStatus;
-import org.apache.coyote.http11.Session;
+import org.apache.catalina.Session;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
@@ -17,7 +20,9 @@ import com.techcourse.service.UserService;
 
 public class LoginController extends Controller {
     private static final LoginController instance = new LoginController();
+    private static final SessionManager sessionManager = SessionManager.getInstance();
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
+    private static final String SESSION_ATTRIBUTE = "user";
 
     private final UserService userService = new UserService();
 
@@ -49,13 +54,10 @@ public class LoginController extends Controller {
         User user = userService.login(account, password);
         log.info("User found: {}", user);
 
-        HttpCookie httpCookie = new HttpCookie(request.getCookie());
-        if (httpCookie.hasJSessionId()) {
-            return redirect("index.html", new HttpResponse());
-        }
         HttpResponse response = new HttpResponse();
-        Session session = Session.createRandomSession();
-        session.setAttribute("user", user.getAccount());
+        Session session = sessionManager.findSession(request).orElse(Session.createRandomSession());
+        session.setAttribute(SESSION_ATTRIBUTE, user.getAccount());
+        sessionManager.add(session);
         response.setCookie(HttpCookie.ofJSessionId(session.getId()));
 
         return redirect("index.html", response);
@@ -64,8 +66,8 @@ public class LoginController extends Controller {
 
     @Override
     protected HttpResponse doGet(HttpRequest request) throws IOException {
-        HttpCookie httpCookie = new HttpCookie(request.getCookie());
-        if (httpCookie.hasJSessionId()) {
+        Optional<Session> session = sessionManager.findSession(request);
+        if (session.isPresent() && Objects.nonNull(session.get().getAttribute(SESSION_ATTRIBUTE))) {
             return redirect("index.html", new HttpResponse());
         }
         return redirect("login.html");
